Refactored project and added folder structure
